data_part/requirements.txt
openeo
h5netcdf
matplotlib
pyarrow 
rioxarray 
geopandas 
openeo[jupyter]
staticmap
pillow
requests
geotiler
ultralytics
torch
numpy
tqdm
pandas